IS-4149: Remove depricated Panel-component and consolidate "ingen møter"-view for both tabs
Endret Panel til å bruke Box, og oppdaget at det ser ulikt ut på de to tabsene når det ikke finnes noen møter. Sjekken for ingen møter skjedde to ganger i enhetens-møter kjeden, så fjernet den et sted, og da sto man igjen med en komponent som ikke gjorde noen ting. Denne ble fjernet.
